Initial commit with python program
ecliptic_banner_chart.py creates a series of SVG files each with a star
chart, including constellation figures of the zodiac and a line marking
the ecliptic, in plate carrée projection and ecliptic coordinates. Each
file in the series is anchored on a different constellation of the
zodiac, providing a selection of images to use for the final product.